requirements.txt
gspread>=5.0.0
gspread-dataframe>=3.0.0
google-auth>=2.0.0
snowflake-connector-python>=2.7.9
pandas>=1.3.0
PyYAML>=5.4.1